Cleanup refactoring
