--- implementation/jobberknoll/core/domain/entities/account-base.ts
@@ -1,10 +1,0 @@
-import type { UUID } from "@jobberknoll/core/shared";
-
-export type AccountBase = {
-  id: UUID;
-  fullName: string;
-  email: string;
-  hashedPassword: string;
-  isActive: boolean;
-  lastModified: number;
-};
