--- implementation/jobberknoll/core/domain/entities/admin.ts
@@ -1,3 +1,0 @@
-import type { AccountBase } from "./account-base.ts";
-
-export type Admin = AccountBase & { type: "admin" };
