--- implementation/jobberknoll/core/domain/entities/driver.ts
@@ -1,3 +1,0 @@
-import type { AccountBase } from "./account-base.ts";
-
-export type Driver = AccountBase & { type: "driver" };
