--- implementation/jobberknoll/core/domain/entities/inspector.ts
@@ -1,3 +1,0 @@
-import type { AccountBase } from "./account-base.ts";
-
-export type Inspector = AccountBase & { type: "inspector" };
