+++ implementation/jobberknoll/core/domain/entities/mod.ts
@@ -1,5 +1,1 @@
 export * from "./account.ts";
-export * from "./admin.ts";
-export * from "./driver.ts";
-export * from "./inspector.ts";
-export * from "./passenger.ts";
