--- implementation/jobberknoll/core/domain/entities/passenger.ts
@@ -1,6 +1,0 @@
-import type { AccountBase } from "./account-base.ts";
-
-export type Passenger = AccountBase & {
-  type: "passenger";
-  phoneNumber?: string;
-};
